% Generated by roxygen2 (4.0.1): do not edit by hand
\name{asISO8601}
\alias{asISO8601}
\title{Converts an object which can be converted to a POSIXlt object to a ISO8601 date time string.}
\usage{
asISO8601(dt)
}
\description{
Converts an object which can be converted to a POSIXlt object to a ISO8601 date time string.
}

