% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{desemprego}
\alias{desemprego}
\title{Desemprego}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5428 rows and 3 columns.
}
\source{
\url{https://www.imf.org/external/datamapper/datasets/WEO}
}
\usage{
desemprego
}
\description{
\describe{
  \item{desemprego}{porcentagem, variação anual}
}
}
\keyword{datasets}
