% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dividabruta}
\alias{dividabruta}
\title{Dívida Bruta do Governo}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10212 rows and 3 columns.
}
\source{
\url{https://www.imf.org/external/datamapper/datasets/WEO}
}
\usage{
dividabruta
}
\description{
\describe{
  \item{dividabruta}{porcentagem do pib}
}
}
\keyword{datasets}
