% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dividaliquida}
\alias{dividaliquida}
\title{Dívida Líquida do Governo}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10350 rows and 3 columns.
}
\source{
\url{https://www.imf.org/external/datamapper/datasets/WEO}
}
\usage{
dividaliquida
}
\description{
\describe{
  \item{dividaliquida}{porcentagem do pib, variação anual}
}
}
\keyword{datasets}
