% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMI.R
\name{FMI}
\alias{FMI}
\title{On the conventional definition of path-specific effects - fully mediated interaction with multiple ordered mediators.}
\usage{
FMI(
  data,
  outcome = "outcome.name",
  treat = "treat.name",
  mediators,
  baseline.conf = c("BaselineConf.name1", "BaselineConf.name2", ...),
  med.model,
  outcome.model,
  a0 = 0,
  a1 = 1,
  MonteCarlo.time = 10^4,
  seed = 1234,
  nboost = 1000
)
}
\arguments{
\item{data}{A data frame, where the column is the variable and row is the sample.}

\item{outcome}{Variable name of the outcome.}

\item{mediators}{A vector of ordered mediator names.}

\item{baseline.conf}{A vector of confounder names.}

\item{med.model}{A vector of model names (GLM) to specify the mediator models.}

\item{outcome.model}{The outcome model}

\item{exposure}{Variable name of the treatment or exposure.}
}
\value{
Estimated PSEs, FMI, and total effect (TE)
}
\description{
FMI calculates the causal path-specific effects by decomposing fully mediated interaction.
}
\examples{
FMI(data=sim.data,outcome="feeltherm",treat = "spifev1per",mediators=c("m1_crqdysmnP","m2_hadsanxN"),
baseline.conf = c("countryB","female","ageT0","bmi","smopackyrs"),
med.model=c("gaussian","gaussian"),
outcome.model = "gaussian",nboost=100)
}
\author{
An-Shun Tai \email{anshuntai@nctu.edu.tw}
}
