% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{fmi_lightnings}
\alias{fmi_lightnings}
\title{Get data on lightning strikes within a given time frame.}
\usage{
fmi_lightnings(start, end, bbox, crs = NULL)
}
\arguments{
\item{start}{Character start date "YYYY-MM-DD" for the query.}

\item{end}{Character end date "YYYY-MM-DD" for the query.}

\item{bbox}{Extent object defining the bounding box for the query.}

\item{crs}{CRS object used to project the results from EPSG:4326 (optional).}
}
\value{
SpatialPointsDataFrame object. The attribute table of the returned
  object contain the following fields:

\itemize{
  \item \code{cloud_indicator}: is the lighting a ground (0) or cloud (1)
    lightning?
  \item \code{ellipse_major}: ???
  \item \code{multiplicity}: ???
  \item \code{peak_current}: ???
}
}
\description{
Thin wrapper around \code{FMIWFSClient$getLightningStrikes()}. Accessing 
the \code{FMIWFSClient} is done through the \code{.session} environment
and the associated functions.
}
\seealso{
\link{\code{FMIWFSClient$getLightningStrikes()}}
}

