% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fmi_weather_stations}
\alias{fmi_weather_stations}
\title{Get all active FMI weather stations.}
\usage{
fmi_weather_stations()
}
\value{
dataframe of active weather stations
}
\description{
Table of active weather stations has been manually copied from FMI web 
pages and is not fetched over the API. Table is provided as a csv file
within \code{fmi} package.
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}
\seealso{
\url{http://ilmatieteenlaitos.fi/havaintoasemat?p_p_id=stationlistingportlet_WAR_fmiwwwweatherportlets&p_p_lifecycle=0&p_p_state=normal&p_p_mode=view&p_p_col_id=column-4&p_p_col_count=1&_stationlistingportlet_WAR_fmiwwwweatherportlets_stationGroup=WEATHER}
}

