% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{FMIWFSClient-class}
\alias{FMIWFSClient}
\alias{FMIWFSClient-class}
\title{A class to make requests to the FMI open data API.}
\description{
A class to make requests to the FMI open data API.
}
\section{Methods}{

\describe{
\item{\code{getRasterURL(request, parameters)}}{Returns raster URL from a WFS query response.}
}}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\references{
See citation("fmi")
}

