% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{FMIWFSFileClient-class}
\alias{FMIWFSFileClient}
\alias{FMIWFSFileClient-class}
\title{A class to make requests to the FMI open data API.}
\description{
A class to make requests to the FMI open data API.
}
\author{
Jussi Jousimo \email{louhos@googlegroups.com}
}
\references{
See citation("fmi")
}

