% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{FMIWFSRequest-class}
\alias{FMIWFSRequest}
\alias{FMIWFSRequest-class}
\title{A class to build WFS request URL to the FMI API.}
\description{
A class to build WFS request URL to the FMI API.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\references{
See citation("fmi")
}

