% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdpcurrent}
\alias{gdpcurrent}
\title{PIB a preços correntes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10442 rows and 3 columns.
}
\source{
\url{https://www.imf.org/external/datamapper/datasets/WEO}
}
\usage{
gdpcurrent
}
\description{
Gross domestic product is the most commonly used single measure
 of a country's overall economic activity. It represents the total
 value at current prices of final goods and services produced within
 a country during a specified time period, such as one year.
 Source: World Economic Outlook
\describe{
  \item{gdpcurrent}{bilhões de U.S. dólares}
}
}
\keyword{datasets}
