% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdppercapita}
\alias{gdppercapita}
\title{PIB per capita, a preços correntes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10396 rows and 3 columns.
}
\source{
\url{https://www.imf.org/external/datamapper/datasets/WEO}
}
\usage{
gdppercapita
}
\description{
Gross domestic product is the most commonly used single measure
of a country's overall economic activity. It represents the total
value at current prices of final goods and services produced within
a country during a specified time period divided by the average
population for the same one year.
Source: World Economic Outlook
\describe{
  \item{value}{dólares per capita}
}
}
\keyword{datasets}
