% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{GeoStatFiWFSRequest-class}
\alias{GeoStatFiWFSRequest}
\alias{GeoStatFiWFSRequest-class}
\title{Builds WFS request to the stat.fi geospatial API.}
\description{
Builds WFS request to the stat.fi geospatial API.
}
\examples{
# See the vignette.
}
\author{
Jussi Jousimo \email{louhos@googlegroups.com}
}
\references{
See citation("fmi")
}

