% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getFinlandBBox}
\alias{getFinlandBBox}
\title{Returns bounding box surrounding Finland}
\usage{
getFinlandBBox()
}
\value{
Bounding box in WGS84 coordinate system as an \code{\link[raster]{extent}} object.
}
\description{
TBA
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
