% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRasterLayerNames}
\alias{getRasterLayerNames}
\title{Return available rasters layer names}
\usage{
getRasterLayerNames(startDateTime, endDateTime, by, variables,
  dateTimeFormat = "\%Y-\%m-\%d")
}
\arguments{
\item{startDateTime}{String start date.}

\item{endDateTime}{String end date.}

\item{by}{TBA.}

\item{variables}{TBA.}

\item{dateTimeFormat}{TBA.}
}
\value{
character vector of raster layer names.
}
\description{
Query is filtered with start and end dates.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}

