% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{init_session}
\alias{init_session}
\title{Initialize a API session.}
\usage{
init_session(apikey)
}
\arguments{
\item{apikey}{Character string valid key for the FMI API.}
}
\description{
Creates a new session based on a provided API key. Function creates a new
\code{\link{FMIWFSClient}} object and assigns it to a environment called
\code{client}.
}
\examples{
# See the vignette.
}
\seealso{
\code{\link[rwfs]{WFSClient}}, \code{\link[rwfs]{WFSCachingClient}}
}
\author{
Joona Lehtomaki  \email{joona.lehtomaki@gmail.com}
}
