% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LongToWideFormat}
\alias{LongToWideFormat}
\title{Convert data from long to wide format.}
\usage{
LongToWideFormat(layer)
}
\arguments{
\item{layer}{Spatial* object.}
}
\value{
layer object.
}
\description{
FIXME: this function is now somewhat specific to BsWfsElement
format. Function could be made more generic, or some sort of class structure
for different response data types needs to be implemented.
}
\note{
response field names \code{fid}, \code{gml_id}, \code{ParameterName}
      and \code{ParameterValue} are hard coded. If these change, the
      function must be adjusted accordingly.
}
\author{
Joona Lehtomaki \email{joona.lehtomaki@gmail.com}
}

