% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transacoescorrentes_gdp}
\alias{transacoescorrentes_gdp}
\title{Saldo em Transações Correntes, \% do PIB}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10350 rows and 3 columns.
}
\source{
\url{https://www.imf.org/external/datamapper/datasets/WEO}
}
\usage{
transacoescorrentes_gdp
}
\description{
\describe{
  \item{transacoescorrentes_gdp}{porcentagem do pib}
}
}
\keyword{datasets}
