% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transformTimeValuePairData}
\alias{transformTimeValuePairData}
\title{Handle and transform TimeValuePairData}
\usage{
transformTimeValuePairData(layer,
  measurementColumnNamePattern = "^result_MeasurementTimeseries_point_MeasurementTVP_value\\\\d*$",
  variableColumnNames, measurementColumnName = "measurement")
}
\arguments{
\item{layer}{XXX object}

\item{measurementColumnNamePattern}{String pattern used to match the 
measurement column.}

\item{variableColumnNames}{String vector used to match the 
variable columns.}

\item{measurementColumnName}{String name for the measurument column.}
}
\value{
layer object.
}
\description{
Response data is massaged into suitable local format.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}

