% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{WFSClient-class}
\alias{WFSClient}
\alias{WFSClient-class}
\title{An abstract class to make requests to a WFS.}
\description{
An abstract class to make requests to a WFS.
}
\author{
Jussi Jousimo \email{louhos@googlegroups.com}
}
\references{
See citation("fmi")
}

