% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{WFSFileClient-class}
\alias{WFSFileClient}
\alias{WFSFileClient-class}
\title{A class to cache response from a WFS to a file.}
\description{
A class to cache response from a WFS to a file.
}
\author{
Jussi Jousimo \email{louhos@googlegroups.com}
}
\references{
See citation("fmi")
}

