% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{WFSStreamClient-class}
\alias{WFSStreamClient}
\alias{WFSStreamClient-class}
\title{A class to stream response from a WFS.}
\description{
A class to stream response from a WFS.
}
\author{
Jussi Jousimo \email{louhos@googlegroups.com}
}
\references{
See citation("fmi")
}

