% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wideToLongFormat}
\alias{wideToLongFormat}
\title{Convert data from long to wide format}
\usage{
wideToLongFormat(layer, timeColumnNamePattern = "^time\\\\d*$",
  measurementColumnNamePattern = "^measurement\\\\d*$",
  variableColumnName = "variable")
}
\arguments{
\item{layer}{XXX object.}

\item{timeColumnNamePattern}{String pattern used to match the 
time column.}

\item{measurementColumnNamePattern}{String pattern used to match the 
measurement column.}

\item{variableColumnName}{String name for the variable column name.}
}
\value{
layer object.
}
\description{
Convert data from long to wide format
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}

