% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fmi}
\alias{fmi}
\title{FMI - World Economic Outlook}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10442 rows and 12 columns.
}
\usage{
fmi
}
\description{
Agregado de todos os dados disponíveis no pacote "fmi".
Crescimento do PIB; PIB a preços correntes; PIB per capita, a preços correntes;
PIB por Paridade do Poder de Compra, preços correntes; Inflação; Desemprego;
Saldo em Transações Correntes; Saldo em Transações Correntes, \% do PIB;
Dívida Líquida do Governo; Dívida Bruta do Governo
}
\keyword{datasets}
