% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{FMIWFSClient-class}
\alias{FMIWFSClient}
\alias{FMIWFSClient-class}
\title{A class to make requests to the FMI open data API.}
\description{
A class to make requests to the FMI open data API.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\references{
See citation("fmi")
}

