% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMIWFSClient.R
\docType{data}
\name{FMIWFSClient}
\alias{FMIWFSClient}
\title{A class to retrieve and manipulate data from the FMI open API}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FMIWFSClient
}
\description{
TBA
}
\section{Methods}{

\itemize{
 \item \code{getDailyWeather}: Returns daily weather time-series
 \item \code{getMonthlyWeatherRaster}: Returns monthly weather raster
}
}
\examples{
# See the vignette.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\references{
See citation("fmi")
}
\seealso{
\code{\link[rwfs]{WFSClient}}, \code{\link[rwfs]{WFSCachingClient}}
}
\keyword{datasets}

