% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{FMIWFSClient}
\alias{FMIWFSClient}
\title{A class to make requests to the FMI open data API.}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x4af4e20> 
 - attr(*, "name")= chr "FMIWFSClient_generator"
}}
\usage{
FMIWFSClient
}
\description{
A class to make requests to the FMI open data API.
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\references{
See citation("fmi")
}
\keyword{datasets}

