% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{FMIWFSRequest-class}
\alias{FMIWFSRequest}
\alias{FMIWFSRequest-class}
\title{A class to build WFS request URL to the FMI API.}
\description{
A class to build WFS request URL to the FMI API.
}
\section{Methods}{

\describe{
\item{\code{getURL()}}{Returns WFS request URL.}
}}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\references{
See citation("fmi")
}

