% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMIWFSRequest.R
\docType{data}
\name{FMIWFSRequest}
\alias{FMIWFSRequest}
\title{A class to build WFS request URL to the FMI API}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FMIWFSRequest
}
\description{
TBA
}
\examples{
\dontrun{request <- FMIWFSRequest$new(apiKey=apiKey)}
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\references{
See citation("fmi")
}
\seealso{
\code{\link[rwfs]{WFSRequest}}, \code{\link[rwfs]{WFSCachingRequest}}
}
\keyword{datasets}

