% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{FMIWFSRequest}
\alias{FMIWFSRequest}
\title{A class to build WFS request URL to the FMI API.}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x49eaa88> 
 - attr(*, "name")= chr "FMIWFSRequest_generator"
}}
\usage{
FMIWFSRequest
}
\description{
A class to build WFS request URL to the FMI API.
}
\examples{
\dontrun{request <- FMIWFSRequest(apiKey=apiKey)}
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}
\references{
See citation("fmi")
}
\keyword{datasets}

