% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFinlandBBox}
\alias{getFinlandBBox}
\title{Returns bounding box surrounding Finland}
\usage{
getFinlandBBox()
}
\value{
Bounding box in WGS84 coordinate system as an \code{\link[raster]{extent}} object.
}
\description{
Returns bounding box surrounding Finland
}
\author{
Jussi Jousimo \email{jvj@iki.fi}
}

